Added pvp.starter integration
If the player doesn't have the pvp.starter permission it forces them to return to spawn and tells them so, else it continues to serialise the inventory and give the kit to the player as normal. Command to deny player entry to starter to follow.
Has not yet been bug tested.